Move WorldState to own file
